adds basic api request for users
