+++ src/components/Users.vue
@@ -5,7 +5,7 @@
         <div class="col-12">
           <card>
             <template slot="header">
-              <h4 class="card-title">Striped Table with Hover</h4>
+              <h4 class="card-title">Users</h4>
               <p class="card-category">Here is a subtitle for this table</p>
             </template>
             <div class="table-responsive">
@@ -13,22 +13,6 @@
                 class="table-hover table-striped"
                 :columns="table1.columns"
                 :data="table1.data"/>
-            </div>
-          </card>
-
-        </div>
-
-        <div class="col-12">
-          <card class="card-plain">
-            <template slot="header">
-              <h4 class="card-title">Table on Plain Background</h4>
-              <p class="card-category">Here is a subtitle for this table</p>
-            </template>
-            <div class="table-responsive">
-              <l-table
-                class="table-hover"
-                :columns="table2.columns"
-                :data="table2.data"/>
             </div>
           </card>
         </div>
@@ -41,44 +25,8 @@
 import LTable from "@/components/UIComponents/Table";
 import Card from "@/components/UIComponents/Cards/Card";
 
-const tableColumns = ["Id", "Name", "Salary", "Country", "City"];
-const tableData = [
-  {
-    id: 1,
-    name: "Dakota Rice",
-    salary: "$36.738",
-    country: "Niger",
-    city: "Oud-Turnhout"
-  },
-  {
-    id: 2,
-    name: "Minerva Hooper",
-    salary: "$23,789",
-    country: "Curaçao",
-    city: "Sinaai-Waas"
-  },
-  {
-    id: 3,
-    name: "Sage Rodriguez",
-    salary: "$56,142",
-    country: "Netherlands",
-    city: "Baileux"
-  },
-  {
-    id: 4,
-    name: "Philip Chaney",
-    salary: "$38,735",
-    country: "Korea, South",
-    city: "Overland Park"
-  },
-  {
-    id: 5,
-    name: "Doris Greene",
-    salary: "$63,542",
-    country: "Malawi",
-    city: "Feldkirchen in Kärnten"
-  }
-];
+const tableColumns = ["Sn", "Uid", "Description", "ObjectClass", "Cn"];
+
 export default {
   name: "Users",
   components: {
@@ -89,13 +37,29 @@ export default {
     return {
       table1: {
         columns: [...tableColumns],
-        data: [...tableData]
-      },
-      table2: {
-        columns: [...tableColumns],
-        data: [...tableData]
+        data: []
       }
     };
+  },
+  created() {
+    this.getUsers();
+  },
+  methods: {
+    getUsers() {
+      fetch("http://localhost:8443/objects/inetOrgPerson/")
+        .then(resp => resp.json()) // Transform the data into json
+        .then(data => {
+          for (var i = 0, len = data.length; i < len; i++) {
+            this.table1.data.push({
+              sn: data[i]["sn"][0],
+              uid: data[i]["uid"][0],
+              description: data[i]["description"][0],
+              objectclass: data[i]["objectClass"][0],
+              cn: data[i]["cn"][0]
+            });
+          }
+        });
+    }
   }
 };
 </script>
